+++ deploy/build.sh
@@ -1,16 +1,15 @@
 #!/bin/bash -e
 
-# Build debian package
-deploy/debian.rb
+export VERSION=$1
 
 # Set up vm
-vagrant up kor
+vagrant up prod
 
 # Extract appliance
 deploy/vagrant.sh appliance
 
 # Tear down vm
-vagrant destroy kor -f
+vagrant destroy -f prod
 
 # Calculate md5 checksums
 deploy/vagrant.sh checksums