+++ deploy/vagrant.sh
@@ -1,4 +1,4 @@
-#!/bin/bash
+#!/bin/bash -e
 
 export DEBIAN_FRONTEND=noninteractive
 
@@ -8,59 +8,63 @@ function system_updates {
   apt-get clean
 }
 
-function install_standalone {
-  debconf-set-selections <<< 'mysql-server mysql-server/root_password password root'
-  debconf-set-selections <<< 'mysql-server mysql-server/root_password_again password root'
+function install_requirements {
+  apt-get install -y \
+    git-core build-essential libmysqlclient-dev libcurl4-openssl-dev ruby-dev \
+    libxml2-dev libxslt-dev imagemagick libav-tools zip
+}
 
+function install_test_requirements {
+  install_requirements
+  apt-get install -y phantomjs
+}
+
+function install_dev_requirements {
+  install_requirements
+  apt-get install -y \
+    libmysqlclient-dev imagemagick libav-tools zip openjdk-7-jre
+}
+
+function install_prod_requirements {
+  install_requirements
+  sudo apt-key adv --keyserver hkp://keyserver.ubuntu.com:80 --recv-keys 561F9B9CAC40B2F7
+  sudo apt-get install -y apt-transport-https ca-certificates
+  sudo sh -c 'echo deb https://oss-binaries.phusionpassenger.com/apt/passenger trusty main > /etc/apt/sources.list.d/passenger.list'
+  sudo apt-get update
+  apt-get install -y apache2 libapache2-mod-passenger
+  sudo a2enmod passenger
+}
+
+function install_elasticsearch {
   wget -qO - http://packages.elasticsearch.org/GPG-KEY-elasticsearch | sudo apt-key add -
   echo 'deb http://packages.elasticsearch.org/elasticsearch/1.5/debian stable main' | tee /etc/apt/sources.list.d/elastic.list
-
   apt-get update
-  apt-get install -y mysql-server elasticsearch
-  apt-get clean
-
+  apt-get install -y openjdk-7-jre elasticsearch
   update-rc.d elasticsearch defaults
+  service elasticsearch start
+}
 
+function install_mysql {
+  debconf-set-selections <<< 'mysql-server mysql-server/root_password password root'
+  debconf-set-selections <<< 'mysql-server mysql-server/root_password_again password root'
+  apt-get install -y mysql-server
   echo "GRANT ALL ON kor.* TO 'kor'@'localhost' IDENTIFIED BY 'kor';" | mysql -u root -proot
 }
 
-function install_requirements {
-  apt-get update
-  apt-get upgrade -y
-  apt-get install -y git-core build-essential ruby apache2 apache2-prefork-dev \
-    libmysqlclient-dev libcurl4-openssl-dev ruby-dev libxml2-dev libxslt-dev \
-    openjdk-7-jre imagemagick ffmpeg libapache2-mod-passenger zip
-  apt-get clean
-}
-
-function install_deb {
-  export VERSION=`cat /vagrant/config/version.txt 2> /dev/null`
-  export DEB_FILENAME="coneda-kor.v$VERSION.deb"
-
-  dpkg -i /vagrant/deploy/build/$DEB_FILENAME
- 
-  su -c "cd /opt/kor/current ; RAILS_ENV=production bundle exec rake db:setup" kor
-  su -c "cd /opt/kor/current ; RAILS_ENV=production bundle exec rake kor:index:create" kor
-  su -c "cd /opt/kor/current ; bundle exec rake assets:precompile" kor
-  su -c "touch /opt/kor/current/tmp/restart.txt" kor
-
-  /etc/init.d/delayed_job start
-}
-
 function appliance {
-  export VERSION=`cat config/version.txt`
+  source /vagrant/tmp/vagrant-env.sh || true
   export OVA_FILENAME="deploy/build/coneda-kor.v$VERSION.ova"
 
-  vagrant halt kor
+  vagrant halt prod
 
-  VBoxManage sharedfolder remove "kor" --name "vagrant"
+  VBoxManage sharedfolder remove "kor.prod" --name "vagrant"
 
-  VBoxManage export kor \
+  VBoxManage export kor.prod \
     --vsys 0 \
     --product "ConedaKOR" \
-    --producturl "http://coneda.net" \
+    --producturl "https://coneda.net" \
     --vendor "Coneda UG" \
-    --vendorurl "http://coneda.net" \
+    --vendorurl "https://coneda.net" \
     --version "$VERSION" \
     --options manifest \
     --output $OVA_FILENAME
@@ -76,8 +80,75 @@ function checksums {
   done   
 }
 
-function rbenv {
-  
+function install_rbenv {
+  apt-get install -y autoconf bison build-essential libssl-dev \
+    libyaml-dev libreadline6-dev zlib1g-dev libncurses5-dev \
+    libffi-dev libgdbm3 libgdbm-dev git-core
+
+  git clone https://github.com/sstephenson/rbenv.git /opt/rbenv
+  git clone https://github.com/rbenv/ruby-build.git /opt/rbenv/plugins/ruby-build
+
+  echo 'export RBENV_ROOT="/opt/rbenv"' >> /etc/profile.d/rbenv.sh
+  echo 'export PATH="/opt/rbenv/bin:$PATH"' >> /etc/profile.d/rbenv.sh
+  echo 'eval "$(rbenv init -)"' >> /etc/profile.d/rbenv.sh
+
+  source /etc/profile.d/rbenv.sh
+
+  RUBY_VERSION=`cat /vagrant/.ruby-version`
+  if [ -d "/opt/kor" ]; then
+    RUBY_VERSION=`cat /opt/kor/.ruby-version`
+  fi
+  rbenv install $RUBY_VERSION
+  rbenv shell $RUBY_VERSION
+  gem install bundler
+
+  chgrp -R vagrant /opt/rbenv/shims
+  chmod -R g+w /opt/rbenv/shims
+}
+
+function clean {
+  apt-get clean
+}
+
+function configure_dev {
+  cd /vagrant
+  bundle install
+  rbenv rehash
+  cp config/database.yml.example config/database.yml
+  bundle exec rake db:drop db:setup
+  bundle exec rake db:test:load
+}
+
+function install_prod {
+  sudo git clone /vagrant /opt/kor
+  sudo chown -R vagrant. /opt/kor
+  cd /opt/kor
+  git checkout v$VERSION
+}
+
+function configure_prod {
+  export RAILS_ENV=production
+  cd /opt/kor
+
+  sudo cp /vagrant/deploy/templates/delayed_job.upstart /etc/init/kor-bg.conf
+  sudo cp /vagrant/deploy/templates/apache.conf /etc/apache2/sites-available/001-kor.conf
+  sudo a2dissite 000-default
+  sudo a2ensite 001-kor
+  sudo service apache2 restart
+
+  sudo cp /vagrant/deploy/templates/gemrc /etc/gemrc
+  sudo cp /vagrant/deploy/templates/logrotate.conf /etc/logrotate.d/kor.conf
+  crontab /vagrant/deploy/templates/crontab
+
+  bundle install --without development test --path /opt/kor/bundle
+  rbenv rehash
+  cp /vagrant/config/database.yml.example config/database.yml
+
+  bundle exec rake db:drop db:setup
+  bundle exec rake assets:precompile
+  bundle exec bin/kor index-all
+
+  sudo service kor-bg start
 }
 
 $1